requirements.txt
kurigram
telethon
TgCrypto
motor
pymongo
python-dotenv
speedtest-cli